New note funcitonality, save notes to SQLite DB, show list of notes on main screen.
Save note on the new note screen takes the note and inserts it into the DB then goes back to the main screen which is refreshed. When main screen is loaded it queries the DB to find all notes and using a card view and recycler view displays a card for each note with the Title, date, time and an image.